# API-Spec.yaml
openapi: 3.0.3
info:
  title: Estimate Generator API
  description: Field Service Management platform for independent contractors and small contracting businesses
  version: 1.0.0
  contact:
    name: Estimate Generator
    url: https://estimategeneratorapp.com
servers:
  - url: https://estimategeneratorapp.com/api/v1
    description: Production server
security:
  - GoogleOAuth: []
  - BearerAuth: []
paths:
...
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: Get customer by ID
      security:
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Customers
      summary: Update customer
      security:
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomer'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Customers
      summary: Delete customer
      security:
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Customer deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /customers/{customer_id}/notes:
    get:
      tags:
        - Customers
      summary: Get customer notes
      security:
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerNote'

    post:
      tags:
        - Customers
      summary: Add note to customer
      security:
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - note
              properties:
                note:
                  type: string
                  maxLength: 1000
      responses:
        '201':
          description: Note added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerNote'

  # Jobs
...
  /jobs/{job_id}:
    get:
      tags:
        - Jobs
      summary: Get job by ID
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'

    put:
      tags:
        - Jobs
      summary: Update job
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJob'
      responses:
        '200':
          description: Job updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'

    delete:
      tags:
        - Jobs
      summary: Delete job
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Job deleted

  # Estimates
...
  /estimates/{estimate_id}:
    delete:
      tags:
        - Estimates
      summary: Delete estimate
      security:
        - BearerAuth: []
      parameters:
        - name: estimate_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Estimate deleted
          
  # Work Orders
...
components:
  securitySchemes:
    GoogleOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect
            email: User email address
            profile: User profile information

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Customer:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip_code:
          type: string
        credit_balance:
          type: number
          format: decimal
          default: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCustomer:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip_code:
          type: string
        credit_balance:
          type: number
          format: decimal
          default: 0
    UpdateCustomer:
      allOf:
        - $ref: '#/components/schemas/CreateCustomer'
    CustomerNote:
      type: object
      required:
        - id
        - customer_id
        - note
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        note:
          type: string
        created_at:
          type: string
          format: date-time

    # Job
    Job:
      type: object
      required:
        - id
        - customer_id
        - title
        - status
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        site_address:
          type: string
        status:
          type: string
          enum: [pending, active, completed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        customer:
          $ref: '#/components/schemas/Customer'
    CreateJob:
      type: object
      required:
        - customer_id
        - title
      properties:
        customer_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        site_address:
          type: string
    UpdateJob:
      allOf:
        - $ref: '#/components/schemas/CreateJob'
        - type: object
          properties:
            status:
              type: string
              enum: [pending, active, completed]

    # Line Item
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...